requirements.txt
pytest==4.5.0
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.6.1-cp38-cp38-win32.whl
pytest-allure-adaptor